+++ app/src/main/java/com/example/weatherapp/models/Weather.java
@@ -1,16 +1,16 @@
 package com.example.weatherapp.models;
 
 public class Weather {
-    private Integer id;
+    private double id;
     private String main;
     private String description;
     private String icon;
 
-    public Integer getId() {
+    public double getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(double id) {
         this.id = id;
     }
